+++ src/index.ts
@@ -4,17 +4,22 @@ type json = {
     [k: string]: any;
 };
 
-export interface FerbyTechLoggerOptions {
+export interface FTLFileOptions {
+    dir: string;
+    logName: string;
+    dailyRotate?: boolean;
+}
+
+export interface FTLOptions {
     console: boolean;
-    file?: {
-        dir: string;
-        logName: string;
-    };
+    consoleColors?: boolean;
+    file?: FTLFileOptions;
     timestamp?: boolean;
 }
 
-export default class FerbyTechLogger {
+export class FerbyTechLogger {
     private history: json[] = [];
+    private cachedDate: Date = new Date();
     private recordHistoryFlag: boolean = false;
     private writeStream: fs.WriteStream | undefined;
     private logLevels = ["debug", "info", "warn", "error"];
@@ -24,16 +29,40 @@ export default class FerbyTechLogger {
         3: [this.logLevels[2], this.logLevels[3]],
         4: [this.logLevels[3]],
     };
+    private colorsByLevel = {
+        debug: "\x1b[36m%s\x1b[0m",
+        info: "\x1b[32m%s\x1b[0m",
+        warn: "\x1b[33m%s\x1b[0m",
+        error: "\x1b[31m%s\x1b[0m",
+    };
 
-    constructor(private readonly options: FerbyTechLoggerOptions) {
-        if (options.file) {
-            /* istanbul ignore else */
-            if (!fs.existsSync(options.file.dir)) {
-                fs.mkdirSync(options.file.dir);
+    constructor(private readonly options: FTLOptions) {
+        this.initWriteStream();
+    }
+
+    private initWriteStream(cb?: (writeStream: fs.WriteStream) => void) {
+        if (this.options.file) {
+            const fileOptions = this.options.file;
+
+            /* istanbul ignore next */
+            if (!fs.existsSync(fileOptions.dir)) {
+                fs.mkdirSync(fileOptions.dir);
             }
 
-            this.writeStream = fs.createWriteStream(`${options.file.dir}/${options.file.logName}.json`);
+            const logName = fileOptions.logName;
+            const formattedDate = this.formatDate(this.cachedDate);
+            const fileName = fileOptions.dailyRotate ? `${logName}-${formattedDate}` : logName;
+            this.writeStream = fs.createWriteStream(`${fileOptions.dir}/${fileName}.json`, { flags: "a" });
+            if (cb) cb(this.writeStream);
         }
+    }
+
+    private validateLogLevel(level: string) {
+        if (process.env.LOG_GROUP && (this.logGroups as any)[process.env.LOG_GROUP]) {
+            return (this.logGroups as any)[process.env.LOG_GROUP].includes(level);
+        }
+
+        return !process.env.LOG_LEVEL || process.env.LOG_LEVEL.toLowerCase() === level;
     }
 
     private write(json: json | string, level: string) {
@@ -49,11 +78,24 @@ export default class FerbyTechLogger {
             }
 
             if (this.options.console) {
-                (console as any)[level](json);
+                if (this.options.consoleColors) {
+                    (console as any)[level]((this.colorsByLevel as any)[level], json);
+                } else {
+                    (console as any)[level](json);
+                }
             }
 
             if (this.writeStream) {
-                this.writeStream.write(JSON.stringify(json) + "\n");
+                if (this.shouldRotate()) {
+                    this.cachedDate = new Date();
+                    this.writeStream.end(() => {
+                        this.initWriteStream((writeStream) => {
+                            writeStream.write(JSON.stringify(json) + "\n", "utf-8");
+                        });
+                    });
+                } else {
+                    this.writeStream.write(JSON.stringify(json) + "\n", "utf-8");
+                }
             }
 
             if (this.recordHistoryFlag) {
@@ -62,12 +104,25 @@ export default class FerbyTechLogger {
         }
     }
 
-    validateLogLevel(level: string) {
-        if (process.env.LOG_GROUP && (this.logGroups as any)[process.env.LOG_GROUP]) {
-            return (this.logGroups as any)[process.env.LOG_GROUP].includes(level);
-        }
+    formatDate(date: Date) {
+        let month = `${date.getMonth() + 1}`;
+        let day = `${date.getDate()}`;
 
-        return !process.env.LOG_LEVEL || process.env.LOG_LEVEL.toLowerCase() === level;
+        /* istanbul ignore next */
+        if (month.length < 2) month = "0" + month;
+
+        /* istanbul ignore next */
+        if (day.length < 2) day = "0" + day;
+
+        return `${date.getFullYear()}-${month}-${day}`;
+    }
+
+    getCachedDate() {
+        return this.cachedDate.getDate();
+    }
+
+    shouldRotate() {
+        return this.options.file?.dailyRotate && this.getCachedDate() < new Date().getDate();
     }
 
     getHistory() {
@@ -76,10 +131,12 @@ export default class FerbyTechLogger {
 
     clearHistory() {
         this.history.splice(0, this.history.length);
+        return this;
     }
 
     recordHistory(active: boolean) {
         this.recordHistoryFlag = active;
+        return this;
     }
 
     setLogLevel(level: string) {
@@ -90,6 +147,7 @@ export default class FerbyTechLogger {
         }
 
         process.env.LOG_LEVEL = lvl;
+        return this;
     }
 
     setLogGroup(group: string) {
@@ -98,21 +156,31 @@ export default class FerbyTechLogger {
         }
 
         process.env.LOG_GROUP = group;
+        return this;
     }
 
     debug(json: json | string) {
         this.write(json, this.logLevels[0]);
+        return this;
     }
 
     info(json: json | string) {
         this.write(json, this.logLevels[1]);
+        return this;
     }
 
     warn(json: json | string) {
         this.write(json, this.logLevels[2]);
+        return this;
     }
 
     error(json: json | string) {
         this.write(json, this.logLevels[3]);
+        return this;
+    }
+
+    log(json: json | string) {
+        this.info(json);
+        return this;
     }
 }